Update CloudWatch Agent configuration version
The base image, which is created in spacelift-io/spacelift-worker-image,
now supports (as of the merging of
spacelift-io/spacelift-worker-image#29) versioned upgrades to the
CloudWatch Agent configuration file.

Add support for this functionality, and default to installing the latest
(at time of writing) version.

We also include support for disabling this functionality, for module
users with their own customised base AMI.