Support launch template versioning
This module uses terraform-aws-modules/terraform-aws-autoscaling.

The terraform-aws-autoscaling module supports three input variables
related to launch template versioning:

1. `default_version` - passed directly to `aws_launch_template`
2. `update_default_version` - passed directly to `aws_launch_template`
3. `launch_template_version` - passed to `aws_autoscaling_group` as
   `local.launch_template_version`

It would be helpful to support launch template versioning, especially
the `update_default_version` setting. `update_default_version` is `null`
by default, so if the launch template is updated, the old version is
still selected as "Current."

This commit will add support for the three input variables, and will
update the docs for the new variables.