Ensure autoscaler zip is present on disk
If `var.autoscaler_version` is set, the `null_resource` uses it as the
`keeper` and won't trigger the provisioner again until the version
number changes.

This approach can create issues. On CI runners that start with a clean
workspace for each run (such as on Spacelift worker pools), the file
won't be there on the next run, leading to an archive creation error
from the `archive_file` data source on each subsequent plan or apply:

```text
Error: Archive creation error

  with module.spacelift_worker_pool_ec2.data.archive_file.binary[0],
  on .terraform/modules/spacelift_worker_pool_ec2/autoscaler.tf line 25,
  in data "archive_file" "binary":
  25: data "archive_file" "binary" {

error creating archive: error archiving file: could not archive missing
file: lambda/bootstrap
```

The current approach also pointlessly unzips the archive to
`lambda/bootstrap`, then rezips it back to the same filename that was
downloaded in the first place.

The short-term fix to resolve this error includes:

- Run a `tofu state rm` to remove the `archive_file` address from state
- Set `var.autoscaler_version` back to `"latest"`

This commit will implement a more robust fix:

- Trigger the `null_resource` if `!fileexists()` so that the file will
  be present on disk.
- Instead of unzipping and rezipping the archive, just leave it as-is
  after download and replace `archive_file` with `local_file` that can
  reference the zip directly.