Retry lifecycle manager on missing worker
If an instance is terminated immediately after starting (say for a
unfortunately-timed instance refresh or something), it's very
likely that the instance wouldn't have (yet) had a chance to
register itself with the pool.

When that happens, the lifecycle manager simply logs the "error"
and moves along. This will leave the instance in a 'terminate:wait'
state until the lifecycle hook times out. Given the default behavior
of 'CONTINUE', this will indefinitely leave those instances in the
ASG, but doing nothing.

This will put the SQS message back in the queue, retrying it after
the normal back-off timeout.